Made 'Other' tasks draggable via dnd-kit. Also, refactored code to make OtherTasks a component that is created for each task rather than it containing all NWIP tasks. ContentHolder currently holds all Content for OtherTasks. App.jsx has been refactored to hold all endpoint connection and passing of NWIP tasks. Refactor this soon to move this out of App.jsx.
