+++ client/src/components/ContentHolder.jsx
@@ -1,23 +1,22 @@
-import { DndContext } from '@dnd-kit/core'
-import { useDroppable } from '@dnd-kit/core';
+import { useState, useEffect } from "react";
 
-export default function ContentHolder() {
+import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable'
+import { OtherTasks } from './OtherTasks'
 
-    function droppable(props) {
-        const { setNodeRef } = useDroppable({
-            id: 'droppable-1',
-        });
-    }
+export const ContentHolder = ({ tasks }) => {
+    // Pass array of IDs to SortableContext
+    const taskIds = tasks.map(task => task._id);
 
     return (
-        <>
-            <DndContext>
-                <section>
-                    <div ref={setNodeRef}>
-
-                    </div>
-                </section>
-            </DndContext>
-        </>
-    );
+        <div className='content-holder'>
+            <p class="text-left">Other:</p>
+            <SortableContext items={taskIds} strategy={verticalListSortingStrategy}>
+                <div class="flex flex-col gap-4">
+                    {tasks.map((task) => (
+                        <OtherTasks _id={task._id.toString()} name={task.name} key={task._id} />
+                    ))}
+                </div>
+            </SortableContext>
+        </div>
+    )
 }
